softPlan/api_processos/requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
flake8==3.8.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
mccabe==0.6.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
yapf==0.30.0
zipp==3.1.0
